requirements-dev.txt
# Development dependencies for data_replication project

# Core dependencies
databricks-connect>=13.0.0
databricks-sdk>=0.8.0
databricks-dlt
delta-spark>=2.4.0

# Configuration and validation
pydantic>=2.0.0
pyyaml>=6.0

# CLI and UI
typer>=0.9.0
click>=8.0.0
rich>=13.0.0

# Utilities
requests>=2.28.0
tenacity>=8.0.0
structlog>=22.0.0

# Testing
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Code quality
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0
coverage>=7.0.0
pre-commit>=3.0.0

# Development tools
jupyter>=1.0.0
ipython>=8.0.0
nbformat>=5.7.0
setuptools
wheel
